requirements.txt
numpy==1.21.6
opencv_contrib_python==4.5.5.64
Pillow==9.1.0
scikit_learn==1.0.2
streamlit==1.8.1
